Continue int remplacement

+++ OpenAVRc/translations/tts_hu.cpp
@@ -115,7 +115,7 @@ I18N_PLAY_FUNCTION(hu, playNumber, getvalue_t number, uint8_t unit, uint8_t att)
   }
 }
 
-I18N_PLAY_FUNCTION(hu, playDuration, int seconds PLAY_DURATION_ATT)
+I18N_PLAY_FUNCTION(hu, playDuration, int16_t seconds PLAY_DURATION_ATT)
 {
   if (seconds == 0) {
     PLAY_NUMBER(seconds, 0, 0);
